MM-43 more organizing and renaming
