+++ src/components/Login/Registration/Registration/RegisterPage.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-
 import { useHistory } from "react-router-dom";
 import RegisterForm from "./RegisterForm";
 import Button from "@mui/material/Button";
