+++ src/components/Pages/Forms/EditArtItem.css
@@ -4,7 +4,6 @@
     margin: 20px;
 }
 
-
 /* Label */
 #outlined-textarea-label {
     font-family: 'Cormorant Garamond',
